requirements.txt
fastapi
uvicorn
python-multipart
pydantic
PyJWT
mysql-connector-python
bcrypt
user-agents
python-dotenv
psutil
gputil
torch
torchvision
transformers
peft
pillow
opencv-python
numpy
email-validator
aiofiles
jinja2
httpx
itsdangerous
sqlalchemy
alembic
databases
aiomysql
pytest
pytest-asyncio
httpx[http2]
pytest-coverage
black
flake8
mypy